"9:35am.
```
                initialize = inl s network sample ->
                    inl cd = s.data.cd
                    inl _, size = sample.dim
                    s.data_add {network = init cd network size}
```

Here I am going to need something to store the output, so I might already need to use `infer` here...

9:40am. What dream is right? I am still thinking about what Hinton said about hierarchical RL being like a company optimizing for a very sparse quarterly profit signal and the CEO making subgoals for his subordinates.

That is a perfect analogy of hierarchical RL, but it is one think to say that and quite another to actual burn it into code. If I could do that then I could unify all the insights and literally cause the Singularity. I said a handful of breakthroughs, but if I put down on paper on what I need it would just be hierarchical RL. Presumably hierarchical RL would need various breakthroughs to be made workable.

To start things off, I'd really like it if regular autoencoders became capable of doing the right thing. Either they need to find it in themselves to produce crisp images or the huge instability caused by adversarial training in GANs needs to be resolved.

It is simple, but I do not know how to do this. And I can't have faith into figuring out next state from the previous until a simple task like that becomes mastered.

What dream is right?

I could have a lot more faith if I understood why regular squared error results in blurry images.

9:45am. Can I assume that a better cost function will be found? Or something else will come around?

Can I put my faith in the community and leave it to them to solve this when they fudged exploration?

9:50am. Let me do it like this - I will leave it to the community for a year to deal with this. Speculation needs signposts. For a year, I will leave autoencoders and hierarchical RL completely out of mind. I won't spare even a thought to model-based RL.

In a year I will be done with all this poker stuff using model-free RL.

If in 2020 they are still going nowhere I will give it my own try at putting together the pieces and making autoencoders work better as a step towards model-based RL.

I will believe in them. They have a decent chance of making significant progress given time.

Collectively they are pretty smart and this is an active research area. Though they might miss important points, they are going in the right direction. For now, we are in this together.

9:55am. I will rely on them. While I am trying to put myself together here, they can work on the pieces of the solution.

Is hierarchy just meta-adversity? Adversity with a referee? I won't assume that this is so.

The dream that is right will be the dream of living power. Hierarchical RL means proper credit assignment. And proper credit assignment means not having to live with evolutionary load anymore.

I want to grasp the improvement that lies beyond self improvement. I'll believe that it exists and keep an eye out on principles.

Hierarchical RL implies its existence. I understood that, but since hierarchical RL is still a moonshot, that means that hierarchical self improvement is two moonshots away. If I want to move forward I need to start reaching for it. Conservatism will not get me anywhere."