"12:40pm. That sure was a distraction. To be honest, I am not even sure what I should be doing during the break anymore. I wish I could replace social media with something.
Well, enough of that.

Let put in one final word regarding nested conditioning and then I'll replace those references with ResizeArrays.

...The way regular optimization works is that is spreads the credit evenly throughout the network. The hierarchy that comes through non-linearities is more likely to be incidental and there due to stochastic optimization than a consequence of the problem formulation.

The way credit should be assigned is like a sponge. When the sponge is dry, the top will absorb it, but when it is wet the water will fall deeper. To put it another way, the bottom layers are the ones which should change the least and the top ones are those which should absorb most of the incoming information.

This makes sense - you do not want to mess with more basic features that affect all the higher level cognition unless absolutely necessary.

As a thought experiment, this should apply to things like linear nets.

12:50pm. This has to be it. There are just not that many ways to make autoencoders work better than they do now, much like there are not that many alternatives to ensembling for uncertainty estimation. Maybe a different cost function like the one Vapnik introduced would be one way of doing it, but his proposals are not convincing to me. Neither is his senile rambling to make things more 'mathematical'. Deep learning will be plenty 'mathematical' once it is understood.

Hence the only solutions can come from optimization, and if one looks there the only thing that is not a complete shot out of the blue would be to do nested learning of the kind I proposed and which has been to work on some things.

This is where I should expect the break to appear. If I am wrong then I guess somebody else will have to find it and I will scramble to change my mind, but this is the position I will push forth. It is the simplest possible explanation.

1pm. I will assume that this will take care of model-based RL and unsupervised learning which are aspects of hierarchical learning.

The only missing piece in that case before the Singularity can be sparked would be long term credit assignment.

There is a clue on how to tackle that - temporal difference learning. Somehow it will have to be done internally. How long term memories could be learned from short term backups is not something that is at all clear to me and to be honest, I do not have to figure it out at this juncture.

I will assume that TD is the answer, but I will ignore the great challenges involved in its elaboration.

There is one thing I could do to narrow my predictions. Whether it is higher order differentiation or internal TD, all of that will in the end be done using local rules. Schmidhuber's view that intelligence will prove to be simple in the end is something I agree with.

1:05pm. 'Hey me, don't you want to see it, the Spiral of 2022? The Spiral where you did not give up. The Spiral that conquered the nightmares. Spiral the true language. The Spiral of your dreams. The Spiral that is your power.'

1:10pm. It is not important that I program. It is important that I create a coherent and consistent vision of the future and make it my desire.

Let me go do the chores while I dwell on this for a bit."