"`region_create`, `region_remove`, `region_create'`, `alloc_cost`, `alloc_accuracy`, `get`.
These six are what I need at the moment. And rather than just this, I am going to need the capacity to push and pop actual regions...

I've renamed the first three to something that reflects them better. I forgot how regions work by this point, so let me refresh my memory.

Oh, I also forgot the `accuracy` function.

2:40pm. Let me work on `accuracy`.

```
                cost = inl s {label out} ->
                    inl {bck} = s.error {label out} s.data.output.last s.data.cd // TODO: Make sure that the cost functions takes in `out`.
                    s.data.bck_error.add (heap bck)
                accuracy = inl s {label out} ->
                    Error.accuracy {label out} input s // TODO: Make sure that the accuracy function takes in `out`.
```

This should do it.

2:45pm. Now let me do the auxiliaries from the top.

2:55pm.

```
                region_create = inl s -> s.data_add {cd=s.data.cd.RegionMem.create}
                region_create' = inl s ret ->
                    inb cd = s.data.cd.RegionMem.create'
                    ret (s.data_add {cd})
                region_clear = inl s -> s.data.cd.RegionMem.clear
                alloc_cost = inl s -> s.data.cd.CudaTensor.zero {elem_type=float32; dim=()}
                alloc_accuracy = inl s ->
                    inl f _ = s.data.cd.CudaTensor.zero {elem_type=int64; dim=()}
                    {accuracy = f (); max_accuracy = f ()}
                get = inl s ->
                    Struct.map s.data.cd.CudaTensor.get
```

This should do it. Let me modify `ResizeArray.add` now. After that I'll modify the cost functions.

3pm. I feel like taking a break, but in times like these I should ask myself what exactly am I going to do during that break. Manga? Anime? Games?

My dearest wish is to draw power from those, not to get consumed by them. Right now they are just a meaningless waste of time, but once I tap into the power of machines, I will be able to draw power from them.

That will be such a wonderful era. I want to live in that time. So I fight. I will do what I set out to do.

There is not that much left. If somebody told me that the Singularity is in 3 years and that all the pieces that I need are already here in embryonic form, would I have any hesitation in doing the work that is needed? Of course not. "