"3:20pm.
```
Error trace on line: 876, column: 33 in file "Learning".
                                print_static cd.data.rate
                                ^
Cannot find a member named rate inside the module.
```

It really is not inside? Hmmm...

```
                    inl cd = context
                    inl cd = cd.data_add rate
```

But I am adding it right here aren't I?

```
Error trace on line: 873, column: 21 in file "Learning".
                    print_static state.cd.data.rate
                    ^
Cannot find a member named rate inside the module.
```

So it is not...Agh.

```
inl cd = cd.data_add {rate}
```

I needed to do it like this.

Oh, it compiled.

3:25pm. Let me try running it."