"4pm.
```
{covariance = 0.0004716485; weight = 0.0001220703}
Starting timing for: iteration 0
The time was 00:00:07.0241458 for: iteration 0
Training: 4.197812E-05
Testing: 0.00121724(125/9984)
Starting timing for: iteration 1
The time was 00:00:06.9593669 for: iteration 1
Training: 2.177839E-05
Testing: 0.0007660478(127/9984)
Starting timing for: iteration 2
The time was 00:00:06.9592183 for: iteration 2
Training: 1.023475E-05
Testing: 0.0003906356(128/9984)
Starting timing for: iteration 3
The time was 00:00:06.9579121 for: iteration 3
Training: 1.241033E-05
Testing: 0.000531473(127/9984)
Starting timing for: iteration 4
The time was 00:00:06.9588649 for: iteration 4
Training: 1.014176E-05
Testing: 0.000269437(128/9984)
```

I've been hoping it would get it right on the first try, but no such luck. This error is no doubt due to the recent changes to accuracy and the cost functions.

Let me remove the back part for the softmax cross entropy. I want to see what it what it does without optimization.

```
{covariance = 0.0004716485; weight = 0.0001220703}
Starting timing for: iteration 0
The time was 00:00:07.0169544 for: iteration 0
Training: 0.004957635
Testing: 0.02928909(27/9984)
Starting timing for: iteration 1
The time was 00:00:06.9584080 for: iteration 1
Training: 0.004918132
Testing: 0.02956541(11/9984)
Starting timing for: iteration 2
The time was 00:00:06.9578361 for: iteration 2
Training: 0.004949038
Testing: 0.02993419(13/9984)
Starting timing for: iteration 3
The time was 00:00:06.9597417 for: iteration 3
Training: 0.004952506
Testing: 0.02970588(17/9984)
Starting timing for: iteration 4
The time was 00:00:06.9586969 for: iteration 4
Training: 0.004988623
Testing: 0.02962989(10/9984)
```

This is just sad. I am definitely doing something wrong.

Ah wait...maybe I am not doing addition across minibatches?

Rather than `store_scalar` don't I have add scalar somewhere?

```
add_store_scalar=inl {from to} -> if threadIdx.x = 0 then to .set (to .get + from)
```

Let me add this thing.

```
        |> s.CudaFun.redo {
            neutral_elem=0; redo=(+); out
            }
```

Do I have some kind of map_out here that I could take advantage here?

Plus I can remove the `neutral_elem` here.

4:15pm.

```
    inl mutable_add_redo {out temp} =
        assert (out.dim = (1 :: ())) "The `out` needs to be 1d singleton tensor."
        s.CudaKernel.redo {
            redo=(+); dim=input.dim
            init = inl i -> temp i .get
            outit = inl i x ->
                inl out = out i
                out .set (out .get + x)
            }
```

Let me use this mutable_add_redo in accuracy and softmax entropy. After that I need to remember to modify the regular functions.

4:20pm.

```
he Spiral Language\Temporary\bin\Release
{covariance = 0.0004716485; weight = 0.0001220703}
Starting timing for: iteration 0
The time was 00:00:07.0280634 for: iteration 0
Training: 126.3198
Testing: 4.740835(9686/9984)
Starting timing for: iteration 1
The time was 00:00:06.9568360 for: iteration 1
Training: 17.73041
Testing: 3.802667(9729/9984)
Starting timing for: iteration 2
The time was 00:00:06.9599732 for: iteration 2
Training: 8.067183
Testing: 3.599667(9746/9984)
Starting timing for: iteration 3
The time was 00:00:06.9577369 for: iteration 3
Training: 4.13315
Testing: 3.26695(9770/9984)
Starting timing for: iteration 4
The time was 00:00:06.9588016 for: iteration 4
Training: 2.357812
Testing: 3.331268(9776/9984)
```

This is almoost right.

I am surprised that I am getting this given that I should be dividing by the outer span.

```
    print_static (to float train_images.basic.span_outer2)
    qwe
    inl cost = agent.get cost / to float train_images.basic.span_outer2
```

Yeah, this should be right.

The division is there.

4:25pm. Ahhh...Why did I do `add_store_scalar`? I should not have done that.

```
Starting timing for: iteration 0
The time was 00:00:07.0185253 for: iteration 0
Training: 0.3528581
Testing: 0.1029162(9684/9984)
Starting timing for: iteration 1
The time was 00:00:06.9533326 for: iteration 1
Training: 0.06728244
Testing: 0.08097984(9732/9984)
Starting timing for: iteration 2
The time was 00:00:06.9570087 for: iteration 2
Training: 0.03067444
Testing: 0.08011956(9739/9984)
Starting timing for: iteration 3
The time was 00:00:06.9586880 for: iteration 3
Training: 0.01610869
Testing: 0.07267401(9759/9984)
Starting timing for: iteration 4
The time was 00:00:06.9570506 for: iteration 4
Training: 0.008868275
Testing: 0.07269876(9762/9984)
```

Well, here it is. The results are quite similar yet distinct to what I had before. This is probably with weight sampling. Let me remove that.

```
{covariance = 0.0004716485; weight = 0.0001220703}
Starting timing for: iteration 0
The time was 00:00:07.0371557 for: iteration 0
Training: 2.301581
Testing: 2.301131(1133/9984)
Starting timing for: iteration 1
The time was 00:00:06.9918025 for: iteration 1
Training: 2.30129
Testing: 2.301117(1133/9984)
```

This is without weight sampling.

4:30pm.

```
    inl weight_sample {d with stddev weight front back} s =
        match s.data with
        //| {rate} ->
        //    inl random = s.CudaRandom.create {stddev dst=.Normal; mean=0f32} {elem_type=float; dim=weight.basic.dim}

        //    inb a = s.CudaBlas.trmm .Left .Lower .nT .NonUnit 1f32 front.sampling.basic random.basic |> CudaAux.temporary
        //    inb b = s.CudaBlas.trmm .Right .Lower .nT .NonUnit 1f32 back.sampling.basic a |> CudaAux.temporary

        //    { d with weight = View.zip {(View.unzip weight) with primal = (s.CudaBlas.geam' .nT .nT one self.basic one b.basic random; View.wrap weight.dim random)} }
        | _ ->
            d
```

Maybe I'll leave `learning2` for tomorrow. Holy crap this is annoying, but I'll get through it. Usually when I remove weight sampling, it remained working. Right now this behavior is quite inexplicable.

4:35pm. Ok, before I continue, let me just turn on the asserts so I can make sure that the thing still works properly and these differences are there only due to numerical instability.

4:35pm. Aggggggggggggghhhhhhhhhhhhhhhhhhhhh...the reason it does not work without weight sampling is because the init is to zero.

That is one mystery down.

Let me fix the error so it does addition to cost and then...I guess I'll look into `learning2`, but given how long `learning1` took I'd rather leave it for tomorrow.
"