"4:45pm.
```
Error trace on line: 133, column: 17 in file "Learning".
                adj .modify' (inl in x -> in + x) (bck {in}) // The adjoint is assumed to be 1 for cost functions.
                ^
Error trace on line: 478, column: 23 in file "Learning".
            bck = inl {in={x y}} -> {x = two * (x - y); y = -(two * (x - y))}
                      ^
`is_numeric a && get_type a = get_type b` is false.
a={adjoint=var (float32); block=[]; primal=var (float32)}, b=var (float32)
```

Just a bit and I will get it done.

```
{covariance = 0.0004716485; weight = 0.0001220703}
Starting timing for: iteration 0
The time was 00:00:07.0083009 for: iteration 0
Training: 0.4366706
Testing: 0.3850085(9382/9984)
Starting timing for: iteration 1
The time was 00:00:06.9175048 for: iteration 1
Training: 0.3338725
Testing: 0.3150052(9517/9984)
Starting timing for: iteration 2
The time was 00:00:06.9255967 for: iteration 2
Training: 0.2759389
Testing: 0.2744099(9568/9984)
Starting timing for: iteration 3
The time was 00:00:06.9297294 for: iteration 3
Training: 0.2371326
Testing: 0.2378332(9640/9984)
Starting timing for: iteration 4
The time was 00:00:06.9291896 for: iteration 4
Training: 0.2147712
Testing: 0.2195129(9634/9984)
```

This is good enough for untuned squared error. It is fine.

---

With this I am done with `learning1`. The feedforward agent is operational.

4:50pm. The next comes `recurrent`. Surprisingly I actually feel like doing it. I won't be testing it today, but let me at least design it. I had an idea for a better kind of design that what I had last time.

For `truncate` I am going to take advantage of the thing I did now for `final`.

That pattern is a good one when it comes to inference.

4:55pm. Let me give it a shot. I'll start by copy pasting the feedforward thing."