"1:50pm. Where was I? For now let me focus on finishing `learning2`. Remember, there is nothing worth doing more than this. That is the chant that will lead me to victory. Once I am done with deep exploration I will be able to grasp something. I will indulge myself in January 2019 for one final time.
After that I do not care - I will select the absolute best tool to conquer the poker quest once and for all. I do not care that involves going with Spiral or making into a JIT for some other language, or abandoning it entirely in favor of something else. I will attain the pinnacle of expressiveness, and performance with my own two hands.

In 2016 it was just too harrowing not being able to do in a way that should have been possible. But now I need to go beyond. I have absolute ability in a style that very few other people do. I have nothing left to prove. I just need to crack deep learning once and for all with the pieces that have been assembled.

I do not know how the autoencoder saga will go after I am done with exploration, but there is one thing I am sure - the question of blurriness is pivotal. I should direct all my search efforts on cracking it, starting with investigating PILCO and Gaussian Processes.

If I can do that then with both exploration and unsupervised learning cracked even if I cannot really go further than that, the programs themselves will point the way to the next part. And that will be something to look forward to. That point called the Beyond.

2pm. Let me start. Today let me pour all my effort into conquering `learning2`. The poker player will be trivial after that."