"``` {covariance = 0.0004716485; weight = 0.0001220703} Starting timing for: iteration 0 The time was 00:00:05.5878009 for: iteration 0 Training: 0 Starting timing for: iteration 1 The time was 00:00:05.5133446 for: iteration 1 Training: 0 Starting timing for: iteration 2 The time was 00:00:05.5110535 for: iteration 2 Training: 0 Starting timing for: iteration 3 ```
What is going on here?

```
                cost = inl s {label out} ->
                    match s.data.buffer.foldr with
                    | {final} -> final {label out}
                    | _ -> ()
```

Well, I am not sure if this is it, but I definitely should not...wait, `foldr`. There is no way this is not a type error...

Ah, no actually it would not be. It just would never get run.

That explains why I am getting 0.

2:20pm. Replaced it with `last`. Let me try it again.

```
Error trace on line: 955, column: 41 in file "Learning".
                                        match out with
                                        ^
Error trace on line: 956, column: 43 in file "Learning".
                                        | {cost accuracy} ->
                                          ^
Error trace on line: 959, column: 43 in file "Learning".
                                        | {cost} ->
                                          ^
Pattern miss error. The argument is <function>
```

2:25pm. It complies once more. But all the extra compilation that it does now puts the compilation time at 3.1s.

...This is quite a lot. I've tried, but I cannot overcome it within the limits I am operating in.

2:25pm.

```
{covariance = 0.0004716485; weight = 0.0001220703}
Starting timing for: iteration 0
The time was 00:00:05.8345020 for: iteration 0
Training: 2.267047
Starting timing for: iteration 1
The time was 00:00:05.7658020 for: iteration 1
Training: 1.851339
Starting timing for: iteration 2
The time was 00:00:05.7694474 for: iteration 2
Training: 1.771574
Starting timing for: iteration 3
The time was 00:00:05.7673583 for: iteration 3
Training: 1.726947
Starting timing for: iteration 4
The time was 00:00:05.7685325 for: iteration 4
Training: 1.704554
```

With this `learning2` is done.

It takes 3.1s to compile roughly. Let me put in some join points in `infer`.

...Now it takes 3s. Yeah, it is hard.

Let me save here and I'll backtrack to a week ago just so I can compile `learning2`."