"``` Running learning2. Time for parse: 00:00:00.7891261 Time for prepass: 00:00:00.1870028 Time for peval was: 00:00:02.5520262 Time for codegen was: 00:00:00.3387666 ```
Well, it the new thing is worse, but 10-15% only. Let's say it is acceptable in terms of slowdown. If there is anything that is unacceptable it would be that it is taking 2.5 instead of under 0.1.

2:35pm. The code produced is ~7k which roughly what I remember it being last time.

2:40pm. With this I am done with `learning2`.

What comes next? Ah yes, `learning1`. I am going to take out the feedforward agent as it is unsafe."