"5:15pm.
```
    inl td reward_scale discount {reward trace value' value} =
        inm r = r * num reward_scale
```

Ah, reward_scale is going to have to come in the agent. No way around that.

5:20pm.

```
    inl td discount {reward trace value' value} =
        inm error = value' - value
        inm _ = sqr error |> as_cost

        inm trace = sigmoid trace
        inm value' = discount * (trace * value' + (one - trace) * primal value)

        succ {value' error=primal error}
```

Let me do it like this.

This is good. Let me modify `ac_sample` next.

```
inl td {trace value} discount {reward value'} =
```

Actually, let me do it like this. This way will be the most concise.

```
    inl grad_rescale {scale input} =
        inm error = abs (get_adjoint input) - scale
        inm cost = sqr error / num 2 |> as_cost
        grad_rescale {scale input}
```

Is this still inside the library? Let me get rid of this.

5:25pm.

```
inl {out bck=bck'} = map (CudaAD.Activation.td reward_scale) {d with trace value} s
```

Ah, I forgot that it calls map. Nevermind.

5:30pm.

```
        inl ac_sample_action {policy trace value} s =
            inl {out bck} = sampling_pg policy s
            {
            out
            bck=inl discount d ->
                inl {out bck=bck'} = map (CudaAD.Activation.td discount) {d with trace value} s
                inl {value' error} = View.unzip out |> module_map (const View.zip)
                {
                out={reward=dyn zero; value'=value}
                bck=inl _ -> bck' (); bck {reward=error}
                }
            }
```

This seems fine now.

...There is one thing that has me worried. Let me check that `as_cost` does allocation.

```
inl as_cost =
    unary_bind <| inl x ->
        inl out = primal x
        {
        out
        bck=inl _ -> adjoint x 0 <- one
        }
```

Ah, crap no it does not. It really should be though.

...Ah, no wait. It does not have to after all. It only sets the adjoint to 1 on the backward pass. That is fine then.

5:35pm. Everything is wonderful now. I've implemented the RL agent and made the necessary arrangements so I can move to testing tomorrow.

Tomorrow I will get `grid1` out of the way and then come the poker tests. I think I should be done with both tomorrow depending on how much debugging of the RL agent I have to do. I do not think it should be too bad even with all the changes that I had introduced. The difficult one was the recurrent agent and the feedforward that came before it.

When all is said and done, there is a beauty to various kinds of abstractions working together towards the same goal.

5:40pm. This is a great spot to stop for the day so I'll stop here. I did well. I was focused throughout and it shows. With this kind of pace it won't be long at all before I am trough with anchored ensembling and deep exploration."