"5:40pm. Now let me see, what that reply was. I sometimes slip in to the old habit, but it nowhere as bad as it was now that I am focused. It happens much less often and I catch myself in the first second.
5:55pm. Ah, it is from the author of the post saying that he will check it out.

6:05pm. Did a little reply.

https://blog.jle.im/entry/practical-dependent-types-in-haskell-1.html

Somebody also linked to this. I saw this years ago before I started working on Spiral and now that I am looking at it again it is like I am seeing it in a completely different light. Reading this code it now gives me a sense of freedom compared to the dread I felt previously. It feels like I do not have any weights on.

...I understand dependent typing now. Completely. This is despite not touching a language with extensional typing for the past two years apart from F#.

It is like I got an unexpected windfall for my efforts. It is no wonder as dependent typing and intensional typing has a significant amount in common.

6:10pm. I can do this. I just have to keep practicing and I will keep getting better.

I've been wondering whether I should go for a dynamic language, but if I understand dependent typing, I really ...

6:35pm. Done with lunch. Though I am bound to be crushed, but if I really want to test myself it might be interesting to give Coq a try again.

...Let me just take a peek at the Software Foundations book while I wait for Egao no Daika to finish downloading. It should be an interesting experiment to see if I really understand dependent typing as I think I do rather than those small fragments."