"1:30pm. I got the poker player working again. This is a great spot to stop for a bit and reevaluate.
Just looking at the damn thing run is making my mood turn sour. It just works so badly for longer sequences. This is completely unlike the tabular player. What, just what do I need to do to stabilize it?

My internal speech is starting to turn negative. I have to ask myself at this point - am I trying to make myself quit work on Spiral?

1:35pm. Spiral is not really meeting my needs. I am putting a lot of effort into it, but I need to move faster - much faster if I am going to win. I want a language that is a lot better than it. Yeah, I want GPU programming, but I want everything else to be effortless as well.

1:40pm. I feel really alone right now.

What I want is a language with an ecosystem on par with Python's, extensional type safety of Haskell, and intensional power of Spiral.

The amount of effort I put into refactoring Spiral code in the past two years has been insane. Everytime I get a bit closer to the answer, but this is not enough.

Right now I am in late 2016 again. I want a language that will help me a lot more.

1:50pm. I very well think that ensembles will do the trick to bring more stability to the RL agent. And deep exploration will help as well. It is a great idea that is worth trying.

But more, I just need more than this.

The approach to GPU programming that Spiral has now I'd rather replace with JITing at runtime. That will put the speed where it needs to be. The rest can be optimized in a standard automatic fashion. Compared to two years ago, a lot of languages got improved support for GPU programming...

How will I face this reality.

1:55pm. The reality is that even 3 years ago I knew a very important lesson - one that I took from the discrete optimization course. That lesson was that amateurs do runtime optimizations, experts do algorithmic ones. Deep learning is definitely in a stage where algorithmic improvements could bring exponential improvements.

Eventually the riddles will be solved and handcoded algorithms will rule and in such a regime Spiral would reign supreme, but right now that is nowhere near that.

Since I've come as far as I can with Spiral, I think it is time I move to the next stage.

I'll give up on what came before and make a bet that it is not the code that was written that matters, but the effort that was put in.

Two new goals:

- See how far I can get in Coq this time. Just how much did my reasoning develop in the past 2.5 years while I tirelessly worked on Spiral.

- As an experiment, implement anchored ensembling plus TD with uncertainty propagation in Python and PyTorch.

Yeah, I know I looked down on Python, but I kind of get the feeling that I should be able to get a lot more power out of a dynamic language now that my programming skills have matured. Despite everything, I do believe that intensional polymorphism is much preferable to macros.

- Hence investigate ways of mixing different languages together. If I expect Spiral or Haskell or Python are any language to be the best in all areas, I am going to be disappointed. Spiral will be absolutely unrivaled for GPU programming. And highly expressive dynamic languages could take a lot of the burden of making good interfaces. Or alternatively, dependently typed languages when they mature.

It is not enough to be good at one aspect of programming. Even if staged functional programming is a unique hidden style, to reach absolute mastery I as a programmer need excellence in all areas even if the individual tools cannot attain that.

Most importantly, I love Spiral and cannot just abandon it. I will definitely make it evolve at some point.

2:10pm. Let me stop here for a bit so I can do the chores.

Then for the next week or so, I'll give going through the first SF volume a try if possible."