"10:50am. Hell of a night, but I figured out everything. I now understand how I could bring the partial evaluator to same scale factor as an interpreter for a dynamic language.
https://fsprojects.github.io/FSharpx.Collections/PersistentVector.html

This thing has o(log32) for indexing and o(1) for appending and size. I mentioned that this might be worth replacing the tuples, but in fact what I should be using it for is to replace the environment itself. It occurred me during the night if I numbered all the variables by their scope, then I could use that to index into the environment much faster than it could be possible ordinarily.

10:55am. At first I thought of various tests I could do, but this cuts to the root of it all.

I am going to have to change how I handle `open`s - I am going to have to disallow it for nonglobal modules.

11am. I'll also change how uncasing of boxed types work - I will no longer do it automatically. It is just too much pain in the ass to use CSE to propagate information. The way I did it is just stupid. I'll make it a manual operation instead.

After that I will remove CSE from the language entirely. The only reason I added it in the first place is too support uncasing.

In addition to that, I will do destructuring only at join point sites (unless they are in the tail position.)

...No, I am going to have to it on bindings after all unless I can find a way to actually detect when it is in tail position or not. Forget it. It is not much of an overhead to check whether a variable needs unboxing. It is no trouble at all. It is a small constant factor everywhere at worst.

11:10am. Before I start anything, I really want to look more into how activation records are supposed to be allocated. Is that data structure (that I do not know how is implemented at the moment) the only real choice here.

11:30am. Ah, it does not have o(1) append, it is also o(log32). I made a dumb error for some reason.

I forgot to mention, at join points I will pass the environments as they are without renaming them. Rather than doing renaming at every damn binding what I will do is do renaming only for getting the key and that is it. Every binding will have a set of live variables that I will be able to use to this renaming efficiently.

11:40am. Let me take a break here.

I'll mention the persistent vector idea on the PL sub later and see if I get any good replies."