"11:55am. https://boards.4chan.org/pol/thread/199081337
It would be cool if bank runs actually started happening. I actually did not think of this, but inducing bank runs is definitely a vector of attack for those in the shadows.

12:45pm. For some reason the screen just went blank and I had to reset to fix it. The monitor I am using can be wonky sometimes.

At any rate, I am done with breakfast. Let me watch Mini Toji and then I'll make a post on activation records.

12:50pm. https://cdn2.desu-usergeneratedcontent.xyz/vr/image/1545/85/1545854960121.png

(the picture of duke says serious-sam.jpg for some reason)

```
He's manly, macho, confident, powerful, wisecracking and funny, but also self-centered, kind of a meathead, and a bit of an asshole. He's not a complex character, but he has personality. He's an endearing parody of classic action movie heroes, being both funny, but also genuinely cool in some ways, as well as sympathetic; alien bastards stealing our chicks is a serious problem which he is dedicated to solving, as I play, I am invested in this plot point, even if there's little narrative and it's just for flavoring.

And this is good because though there's many level which are a bit silly or just generic action, some levels are serious, and even quite dark and grim, especially the darker lit ones with a gloomy tune, and featuring captured women, begging for death. Duke Nukem himself is the shinning golden beacon of comedic relief to all of this (but not TOO comedic), and is an excellent juxtaposition for the game's more dark and serious moments, in a way, I want to say he is what lets the game achieve it's "sweet and sour flavor".

When it's just him and the Overlord, having their showdown, Duke grimly announces his intent to "Rip off his head and shit down his neck." A threatening taunt before a battle, Duke has plenty of one-liners to bring for combat, but this one's different. As Duke defeats the Overlord, you see him tear the Overlord's head off, then take off his gear, unzip his pants, sit down on the creature's neck stump, and take out his newspaper, as he whistles his theme tune.

It was not just some smack talking or empty words, he said that, he meant it, and he goes through with it, because he is full of bravado and capability, and will go through extreme lengths to humiliate and disrespect his worst enemies. It's hillarious and it's absolutely badass.

It's collorful at times, but also often dark and gray.
The game really straddles the line between dark grit and silly action, or really juxtaposing one with the other.

You're on a dark and grungy space station with a tense and creppy soundtrack sounding in the background. The female crew are all captured in alien lairs and I guess raped savegely for breeding more horrid monstrosities, they outright beg for death.

And here comes a big muscled dude in sunglasses named fucking Duke Nukem, who blows up a series of raging alien monsters with a volley from a rocket machinegun, where he exclaims "Holy shit!", then he zaps a fat and flying one with a raygun, which shrinks it, whereupon Duke Nukem crushes said shrunken alien underneath his boot, like it's a cockroach, praising himself with a "Damn, I'm good.", before he jumpkicks a ventialtion grate and crawls inside, crouching, yet somehow managing to kick to death an aggressive parasite that latched itself to his face (in effect kicking himself in the face while crouched over).
```

I found this thing yesterday while wasting time and thought it would be nice to share. Back in 96' I thought that great games would keep coming forever. Now I understand that I have a responsibility to do something about it myself.

https://www.youtube.com/watch?v=gIq4ymJ1YLw
GRABBAG! The Ultimate Duke Nukem Tribute

I've been listening this on repeat for the past two days. Duke Nukem is just such an iconic figure. The entire franchise had only one game worth mentioning, but that was quite a game.

Once I crack the AI puzzle I can only hope that Sol From Hell will become a similar fixture in the global consciousness.

Let me make a little post on the PL sub."