"1:25pm. https://www.reddit.com/r/ProgrammingLanguages/comments/adtp7l/d_what_are_good_choices_for_interpreter/
"For Spiral I've been using F#'s default immutable map, but it just recently it occured to me that if I numbered all the variables I could use the [persistent vector](https://fsprojects.github.io/FSharpx.Collections/PersistentVector.html) instead. I've been having a lot of compile-time performance issues lately in Spiral, or rather the scale of the programs has grown so much that my amateurish implementation is no longer feasible and I need something better.

I like the persistent vector idea in particular as it would make the numbering analysis really straightforward. Some ideas that I've had like doubly nested lists would require a more complicated analysis which I've yet to grasp and possibly grouping of immediately nested lambdas.

It feels that there should be a lot of options for this sort of thing and that it should be a major subject of debate, but everywhere I've looked I only find associative lists being used. Does anybody have references for this subject? What do existing interpreted languages use under the hood?"

1:30pm. https://www.lua.org/doc/jucs05.pdf
The Implementation of Lua 5.0

http://craftinginterpreters.com/a-virtual-machine.html

I got the Lua paper from crafting interpreters. Let me take a look at it.

1:35pm. "We consider that we have achieved our design and implementation goals. Lua
is a very portable language: it runs on any machine with an ANSI C compiler,
from embedded systems to mainframes. Lua is really lightweight: for instance, on
Linux its stand-alone interpreter, complete with all standard libraries, takes less
than 150 Kbytes; the core is less than 100 Kbytes. Lua is efficient: independent
benchmarks [2, 4] show Lua as one of the fastest language implementations in the
realm of scripting languages (i.e., interpreted and dynamically-typed languages).
We also consider Lua a simple language, being syntactically similar to Pascal
and semantically similar to Scheme, but this is subjective."

Maybe Lua would be a good target of study.

https://www.lua.org/docs.html

1:40pm. I was not aware that Lua has an academic focus. That will make studying it easier.

1:50pm. "Research in non-pure functional languages like Scheme and ML has created a vast body of knowledge about compilation techniques for closures (e.g., [19, 1, 21]).3 However, those works do not try to limit the complexity of the compiler. For instance, just the control flow analysis of Bigloo, an optimizer Scheme compiler[20], is more than ten times larger than the whole Lua implementation: The source for module Cfa of Bigloo 2.6f has 106,350 lines, versus 10,155 lines for the core of Lua 5.0. As explained in Section 2, Lua needs something simpler."

Oh, lol, I had no idea about this.

2pm. Persistent vectors are a pretty good idea. With some memory sharing optimizations which would not be too hard to make, they would indeed be suitable for implementing environments.

2:25pm. https://cacm.acm.org/magazines/2018/11/232214-a-look-at-the-design-of-lua/fulltext

This is interesting.

2:55pm. Ok, enough reading. I am not interesting in putting that much effort into learning about various data structures when persistent vector will suffice. I just need to pick one that meets all the requirements and go for it.

I am not interested in small optimizations.

3pm. What is my plan here exactly?

Well, a part of me does just want to move to Python and forget about Spiral for a while.

A different part of me has come to an acceptance that intensional polymorphism is not suitable for large scale programming which requires dynamism.

Both parts are in agreement that join points are the solution to language interop and inlineables are a better choice than macros for 'metaprogramming'. Compiling to simply typed lambda calculus is what is making things so difficult. Instead of F# as target, if I really want to take Spiral further it might be interesting to target dynamic CLR languages.

It would be really easy to switch between dynamic and STLC in the partial evaluator - this would support all styles of programming.

3:05pm. How about I go for one last Spiral hacking session for the next few weeks or so?

I really want to see if improving the environment will speed up the language by a significant amount - it should.

I want to put in an extra bit of effort to leave the language in a more pristine state before I say my farewell to it. Without a doubt, I will come back to Spiral once I've gotten all that I need from Python.

If I could squish those heavy compilation factors and speed up compilation by 10-100x on those difficult example, I will have accomplished a significant thing.

Let me take a break here for a bit."